Allow hosted clients to handle relaunch themselves
This patch adds a new application description parameter "handlesRelaunch".
When there is a relaunch request (launching an application already running)
we check this property. If the application does not provide it, we keep the
former behavior (just focusing first window we find).

But for applications with handlesRelaunch enabled, we just emit the new
IPC call named View_Relaunch, and we pass the call object as parameter.
In this case it is the application that should process the request doing
whatever is most convenient (launching a new window, plainly focusing an
existing window, ignoring the request, etc).

Open-webOS-DCO-1.0-Signed-off-by: Jose Dapena Paz <jose.dapena@palm.com>

Change-Id: I726d62b724508f79d51208bec4a71d8ff144d003