Initial WebAppManager commit
This is a transitional commit that creates basis for the
WebAppManager repository. The history before this commit is
a stripped down version of Luna Sys Manager with only Web App
Manager related parts left in place. Some renames of files
are done in this commit in effort to retain history of those
files after upcoming commits.

Open-webOS-DCO-1.0-Signed-off-by: Jukka Honkela <jukka.honkela@palm.com>

Change-Id: Ibcd1b2b3ed92261202ace6f672e0af094c50ca0c