Fix INCLUDEPATH to be correct for cross compilation
Commit 589b44769bab8ce196d12e5688635943d095bb21 added a link to the global include
directory /usr/include which is not allowed in cross compilation environments like OE and
will causes compilation problems (in OE the LUNA_STAGING variable is not set and so the
include path will expanded to just /usr/include).

As the change was only intended for the desktop build environment it's reverted from
sysmgr-common.pro and added just for the desktop build configuration. The link to the
correct lib directory for the desktop build is already existend in the desktop
configuration and dropped with this.

Open-webOS-DCO-1.0-Signed-off-by: Simon Busch <morphis@gravedo.de>